V1.1
Fixed issue where missiles first would essentially softlock player. Also made chargebeam a "soft" requirement for all bosses. Doing so makes the logic easier because missile damage does not have to be considered before putting an item on a boss but I left the health attribute for the boss object so that functionality could be implemented in the future. This may present boring seeds where charge beam is either front loaded and you know you have to fight a boss or no bosses required and it's a charge beam hunt